lab_05/pyproject.toml
[tool.poetry]
name = "lab-05"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^11.0.0"
numpy = "^2.1.2"
matplotlib = "^3.9.2"
opencv-python = "^4.10.0.84"
tqdm = "^4.66.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
